@@ -2,30 +2,53 @@
   "name": "topchik",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "dev": "next dev --turbopack",
-    "build": "next build",
+    "dev": "next dev --turbopack -p 5175",
+    "build": "next build --turbopack",
     "build:production": "pnpm build",
     "start": "next start",
-    "lint": "next lint",
-    "format": "eslint src --ext .jsx,.js,.tsx,.ts  --no-error-on-unmatched-pattern --fix && stylelint --fix src/{,*/}*.{scss,css} --allow-empty-input",
-    "storybook": "storybook dev -p 6006",
+    "lint": "eslint . .jsx,.js,.tsx,.ts --no-error-on-unmatched-pattern --fix && stylelint --fix \"**/*.{css,scss}\" --allow-empty-input && next lint --no-cache",
+    "format": "prettier --write .",
+    "sb": "storybook dev -p 6006",
     "build-storybook": "storybook build",
     "test": "vitest"
   },
   "dependencies": {
+    "@fontsource/inter": "^5.2.5",
+    "@fontsource/roboto": "^5.2.5",
+    "@hookform/resolvers": "^5.0.1",
+    "@radix-ui/react-checkbox": "^1.1.4",
+    "@radix-ui/react-dialog": "^1.1.6",
+    "@radix-ui/react-dropdown-menu": "^2.1.6",
+    "@radix-ui/react-label": "^2.1.2",
+    "@radix-ui/react-popover": "^1.1.6",
+    "@radix-ui/react-radio-group": "^1.2.3",
+    "@radix-ui/react-scroll-area": "^1.2.3",
+    "@radix-ui/react-select": "^2.1.6",
+    "@radix-ui/react-slot": "^1.1.2",
+    "@radix-ui/react-tabs": "^1.1.3",
+    "@radix-ui/react-toast": "^1.2.6",
     "@reduxjs/toolkit": "^2.6.1",
-    "next": "15.2.3",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "react-hook-form": "^7.54.2"
+    "async-mutex": "^0.5.0",
+    "clsx": "^2.1.1",
+    "next": "15.3.0",
+    "radix-ui": "^1.1.3",
+    "react": "^19.1.0",
+    "react-day-picker": "8.10.1",
+    "react-dom": "^19.1.0",
+    "react-hook-form": "^7.55.0",
+    "react-redux": "^9.2.0",
+    "zod": "^3.24.2"
   },
   "devDependencies": {
-    "@chromatic-com/storybook": "3",
-    "@eslint/eslintrc": "^3",
-    "@it-incubator/eslint-config": "^1.0.4",
+    "@chromatic-com/storybook": "^3.2.6",
+    "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.24.0",
+    "@hookform/devtools": "^4.4.0",
     "@it-incubator/prettier-config": "^0.1.2",
     "@it-incubator/stylelint-config": "^2.0.0",
+    "@next/eslint-plugin-next": "^15.2.4",
     "@storybook/addon-essentials": "8.6.7",
     "@storybook/addon-onboarding": "8.6.7",
     "@storybook/blocks": "8.6.7",
@@ -34,26 +57,33 @@
     "@storybook/react": "8.6.7",
     "@storybook/test": "8.6.7",
     "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^16.2.0",
-    "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "@vitest/browser": "^3.0.9",
-    "@vitest/coverage-v8": "^3.0.9",
-    "eslint": "^9",
+    "@testing-library/react": "^16.3.0",
+    "@types/node": "^22.14.0",
+    "@types/react": "^19.1.0",
+    "@types/react-dom": "^19.1.1",
+    "@typescript-eslint/eslint-plugin": "^8.29.0",
+    "@typescript-eslint/parser": "^8.29.0",
+    "@vitest/browser": "^3.1.1",
+    "@vitest/coverage-v8": "^3.1.1",
+    "date-fns": "^4.1.0",
+    "eslint": "^9.24.0",
     "eslint-config-next": "15.2.3",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-storybook": "^0.11.6",
     "jsdom": "^26.0.0",
     "playwright": "^1.51.1",
-    "sass": "^1.86.0",
+    "prettier": "3.5.3",
+    "sass": "^1.86.3",
     "storybook": "8.6.7",
-    "stylelint": "^16.16.0",
-    "typescript": "^5",
-    "vitest": "^3.0.9"
+    "stylelint": "^16.18.0",
+    "typescript": "^5.8.3",
+    "vitest": "^3.1.1"
   },
   "eslintConfig": {
     "extends": [
       "plugin:storybook/recommended"
     ]
-  }
+  },
+  "packageManager": "pnpm@10.6.5+sha512.cdf928fca20832cd59ec53826492b7dc25dc524d4370b6b4adbf65803d32efaa6c1c88147c0ae4e8d579a6c9eec715757b50d4fa35eea179d868eada4ed043af"
 }
