@@ -1,95 +1,87 @@
-import Image from "next/image";
-import styles from "./page.module.css";
+'use client'
+import Image from 'next/image'
+import '../styles/index.scss'
+import '@fontsource/inter/400.css'
+import '@fontsource/inter/500.css'
+import '@fontsource/inter/600.css'
+import '@fontsource/inter/700.css'
+import styles from './page.module.css'
+import { store } from '@/shared/stores'
+import { Provider } from 'react-redux'
 
 export default function Home() {
   return (
-    <div className={styles.page}>
-      <main className={styles.main}>
-        <Image
-          className={styles.logo}
-          src="/next.svg"
-          alt="Next.js logo"
-          width={180}
-          height={38}
-          priority
-        />
-        <ol>
-          <li>
-            Get started by editing <code>app/page.tsx</code>.
-          </li>
-          <li>Save and see your changes instantly.</li>
-        </ol>
+    <Provider store={store}>
+      <div className={styles.page}>
+        <main className={styles.main}>
+          <Image
+            priority
+            alt="Next.js logo"
+            className={styles.logo}
+            height={38}
+            src="/next.svg"
+            width={180}
+          />
+          <ol>
+            <li>
+              Get started by editing <code>app/page.tsx</code>.
+            </li>
+            <li>Save and see your changes instantly.</li>
+          </ol>
 
-        <div className={styles.ctas}>
+          <div className={styles.ctas}>
+            <a
+              className={styles.primary}
+              href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
+              rel="noopener noreferrer"
+              target="_blank"
+            >
+              <Image
+                alt="Vercel logomark"
+                className={styles.logo}
+                height={20}
+                src="/vercel.svg"
+                width={20}
+              />
+              Deploy now
+            </a>
+            <a
+              className={styles.secondary}
+              href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
+              rel="noopener noreferrer"
+              target="_blank"
+            >
+              Read our docs
+            </a>
+          </div>
+        </main>
+        <footer className={styles.footer}>
           <a
-            className={styles.primary}
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-            target="_blank"
+            href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
             rel="noopener noreferrer"
+            target="_blank"
           >
-            <Image
-              className={styles.logo}
-              src="/vercel.svg"
-              alt="Vercel logomark"
-              width={20}
-              height={20}
-            />
-            Deploy now
+            <Image aria-hidden alt="File icon" height={16} src="/file.svg" width={16} />
+            Learn
           </a>
           <a
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-            target="_blank"
+            href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
             rel="noopener noreferrer"
-            className={styles.secondary}
+            target="_blank"
           >
-            Read our docs
+            <Image aria-hidden alt="Window icon" height={16} src="/window.svg" width={16} />
+            Examples
           </a>
-        </div>
-      </main>
-      <footer className={styles.footer}>
-        <a
-          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/file.svg"
-            alt="File icon"
-            width={16}
-            height={16}
-          />
-          Learn
-        </a>
-        <a
-          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/window.svg"
-            alt="Window icon"
-            width={16}
-            height={16}
-          />
-          Examples
-        </a>
-        <a
-          href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/globe.svg"
-            alt="Globe icon"
-            width={16}
-            height={16}
-          />
-          Go to nextjs.org →
-        </a>
-      </footer>
-    </div>
-  );
+          <a
+            href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
+            rel="noopener noreferrer"
+            target="_blank"
+          >
+            <Image aria-hidden alt="Globe icon" height={16} src="/globe.svg" width={16} />
+            Go to nextjs.org →
+          </a>
+        </footer>
+      </div>
+    </Provider>
+  )
 }
